EE_EB_All_branches

+++ RecoTreeUtils/python/RecoFastTiming_cff.py
@@ -12,7 +12,7 @@ RecoFastTiming = cms.EDAnalyzer("RecoFastTiming",
     ## cut on the impact parameter (assing track to vertex) in cm
     dzCut           = cms.untracked.double(0.2),
     ## track pt and pz threshold for vtx time reconstrution in GeV
-    ptCut           = cms.untracked.double(1),
+    ptCut           = cms.untracked.double(2),
     pz2Cut          = cms.untracked.double(0),
     ## output file collections option
     saveParticles   = cms.untracked.bool(True),
